requirements.txt
streamlit
anthropic
openai
python-dotenv
requests
uuid